omitarg as cmd
